rag-app-lorawan/requirements.txt
aiohappyeyeballs==2.4.0 
aiohttp==3.10.5 
aiosignal==1.3.1 
annotated-types==0.7.0 
anyio==4.4.0 
attrs==24.2.0 
beautifulsoup4==4.12.3 
certifi==2024.8.30 
charset-normalizer==3.3.2 
click==8.1.7 
colorama==0.4.6 
dataclasses-json==0.6.7 
deprecated==1.2.14 
dirtyjson==1.0.8 
distro==1.9.0 
frozenlist==1.4.1 
fsspec==2024.9.0 
greenlet==3.1.0 
h11==0.14.0 
httpcore==1.0.5 
httpx==0.27.2 
idna==3.9 
jiter==0.5.0 
joblib==1.4.2 
llama-cloud==0.0.17 
llama-index-agent-openai==0.3.1 
llama-index-cli==0.3.1 
llama-index-core==0.11.9 
llama-index-embeddings-openai==0.2.5 
llama-index-indices-managed-llama-cloud==0.3.1 
llama-index-legacy==0.9.48.post3 
llama-index-llms-openai==0.2.7 
llama-index-multi-modal-llms-openai==0.2.1 
llama-index-program-openai==0.2.0 
llama-index-question-gen-openai==0.2.0 
llama-index-readers-file==0.2.1 
llama-index-readers-llama-parse==0.3.0 
llama-index-vector-stores-pinecone==0.2.1 
llama-index==0.11.9 
llama-parse==0.5.5 
markdownify==0.13.1 
marshmallow==3.22.0 
multidict==6.1.0 
mypy-extensions==1.0.0 
nest-asyncio==1.6.0 
networkx==3.3 
nltk==3.9.1 
numpy==1.26.4 
openai==1.45.0 
packaging==24.1 
pandas==2.2.2 
pillow==10.4.0 
pinecone-client==5.0.1 
pinecone-plugin-inference==1.1.0 
pinecone-plugin-interface==0.0.7 
pinecone==5.1.0 
pydantic-core==2.23.3 
pydantic==2.9.1 
pypdf==4.3.1 
python-dateutil==2.9.0.post0 
python-dotenv==1.0.1 
pytz==2024.2 
pyyaml==6.0.2 
regex==2024.9.11 
requests==2.32.3 
six==1.16.0 
sniffio==1.3.1 
soupsieve==2.6 
sqlalchemy[asyncio]==2.0.34 
striprtf==0.0.26 
tenacity==8.5.0 
tiktoken==0.7.0 
tqdm==4.66.5 
typing-extensions==4.12.2 
typing-inspect==0.9.0 
tzdata==2024.1 
urllib3==2.2.3 
wrapt==1.16.0 
yarl==1.11.1 
